Updated our feathers app using this command > feathers generate service
Created a validate-user CUSTOM service (not connected to a database). This is blank now, but will make the update command be a way for the user to click a link in an email message to validate the newly created user.

Since this is a custom service, notice the validate-user.class.js file was created in this case.

> feathers generate service
? What kind of service is it? A custom service
? What is the name of the service? validateUser
? Which path should the service be registered on? /validate-user
? Does the service require authentication? No